+++ api/src/main/java/com/team/config/WebSecurityConfig.java
@@ -1,4 +1,4 @@
-package com.team.security.config;
+package com.team.config;
 
 import com.team.security.jwt.JwtAuthenticationEntryPoint;
 import com.team.security.jwt.TokenProvider;
