+++ service/src/main/java/com/team/auth/CustomUserDetailsService.java
@@ -1,12 +1,12 @@
-package com.team.user;
+package com.team.auth;
 
+import com.team.user.User;
+import com.team.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-
-import java.util.ArrayList;
 
 @Service
 @RequiredArgsConstructor
@@ -16,13 +16,12 @@ public class CustomUserDetailsService implements UserDetailsService {
     @Override
     public UserDetails loadUserByUsername(final String email) throws UsernameNotFoundException {
         return userRepository.findByEmail(email)
-                .map(user -> createUser(email, user))
+                .map(this::createUser)
                 .orElseThrow(() -> new UsernameNotFoundException("존재하지 않는 이메일입니다."));
     }
 
-    private org.springframework.security.core.userdetails.User createUser(String email, User user) {
-        return new org.springframework.security.core.userdetails.User(
-                user.getEmail(), user.getPassword(), new ArrayList<>()
-        );
+    private CustomUserDetail createUser(User user) {
+        return CustomUserDetail.create(user);
     }
+
 }
